+++ BoardGameStore.Infrastructure.EFCore/Repositories/UserRepository.cs
@@ -24,8 +24,11 @@ namespace BoardGameStore.Infrastructure.EFCore.Repositories
         public async Task DeleteUser(int id)
         {
             var user = await _context.Users.FindAsync(id);
-            _context.Users.Remove(user);
-            await _context.SaveChangesAsync();
+            if (user != null)
+            {
+                _context.Users.Remove(user);
+                await _context.SaveChangesAsync();
+            }
         }
 
         public async Task<List<UserModel>> GetAllUsers()
